update - 1
I encountered a problem, where were not able to extract data from nested objects, so I have to update my code. 
I used the

"(
                    typeof d[head.key] === 'string' || typeof d[head.key] === 'number' ? d[head.key] : (head.key === "company" ? d[head.key].name : d[head.key].city + ", " + d[head.key].street + ", " + d[head.key].suite)
                )"

the line where I am checking the type of the main object's head key whether it is a string or a number (to be able to display the id as well) when the key has existed in the main d[head.key] will be rendered out in the cells, but in the case of nested objects I had to dive deep and go even further. 

I only needed specific keys like company name and address (street name and number), so I put another nested condition where I am checking when the key name is the company it will automatically return the keys name and in case of address, we are getting back the relevant information.